[#106] refactor: wineSearchSection과 wineFilterSidebar로 구조 분리
