+++ src/app/wines/page.tsx
@@ -7,9 +7,6 @@ export default function WinesPage() {
     <div className='relative grid min-h-screen grid-cols-[auto_1fr_auto] grid-rows-[auto_4rem_1fr_auto] gap-x-4 gap-y-14 xl:grid-cols-[auto_1fr] xl:gap-x-16'>
       <WineRecommendationSection />
       <ReviewActionButton />
-      {/* <WineSearchBar />
-      <WineFilterSidebar />
-      <WineListContainer /> */}
       <WineSearchSection />
     </div>
   );
